LayoutTests/platform/mac/media/media-source/is-type-supported-vp9-codec-check-expected.txt

EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="bad-parameter"') == 'false') OK

Test invalid number of required parameters:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09"') == 'false') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00"') == 'false') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.41"') == 'false') OK

Test valid default optional parameters:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.41.08"') == 'true') OK

Test invalid number of optional parameters:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01"') == 'false') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09"') == 'false') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16"') == 'false') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.09"') == 'false') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.09.01.00"') == 'false') OK

Test valid required and optional parameters:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.09.01"') == 'true') OK

Test invalid codec type
EXPECTED (MediaSource.isTypeSupported('vp10.00.41.08"') == 'false') OK

Test invalid combination of matrixCoefficients and chromaSubsampling
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.41.08.01.01.01.00.00"') == 'false') OK

Test valid profiles:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.41.08"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.41.08"') == 'true') OK
Test invalid profiles:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.01.41.08"') == 'false') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.03.41.08"') == 'false') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.04.41.08"') == 'false') OK

Test valid levels:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.10.08"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.11.08"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.20.08"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.21.08"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.30.08"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.31.08"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.40.08"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.41.08"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.50.08"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.51.08"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.52.08"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.60.08"') == 'true') OK

Test invalid levels:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.61.08"') == 'false') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.62.08"') == 'false') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.70.08"') == 'false') OK

Test valid bitDepths:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.41.08"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.41.10"') == 'true') OK

Test invalid bitDepths:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.41.12"') == 'false') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.00.41.14"') == 'false') OK

Test valid chromaSubsampling:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.00.09.16.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.09.01"') == 'true') OK

Test invalid chromaSubsampling:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.02.09.16.09.01"') == 'false') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.03.09.16.09.01"') == 'false') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.04.09.16.09.01"') == 'false') OK

Test valid colorPrimaries:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.01.16.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.04.16.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.05.16.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.06.16.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.07.16.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.08.16.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.10.16.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.11.16.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.12.16.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.22.16.09.01"') == 'true') OK

Test invalid colorPrimaries:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.03.16.09.01"') == 'false') OK

Test valid transferCharacteristics:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.01.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.04.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.05.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.06.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.07.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.08.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.09.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.10.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.11.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.12.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.13.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.14.09.01"') == 'true') OK

Test invalid transferCharacteristics:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.03.09.01"') == 'false') OK

Test valid matrixCoefficients:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.01.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.04.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.05.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.06.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.07.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.08.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.09.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.10.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.11.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.12.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.13.01"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.14.01"') == 'true') OK

Test invalid matrixCoefficients:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.09.03"') == 'false') OK

Test valid videoFullRangeFlag:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.01.00"') == 'true') OK
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.01.01"') == 'true') OK

Test invalid videoFullRangeFlag:
EXPECTED (MediaSource.isTypeSupported('video/webm; codecs="vp09.02.10.10.01.09.16.01.02"') == 'false') OK

END OF TEST
